move amazon/ into cli/commands/amazon/ for self-contained CLI module
