+++ assistant/src/cli/commands/amazon/checkout.ts
