+++ assistant/src/cli/commands/amazon/product-details.ts
