+++ assistant/src/cli/commands/amazon/request-extractor.ts
@@ -9,8 +9,8 @@
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
 
-import type { SessionRecording } from "../tools/browser/network-recording-types.js";
-import { getDataDir } from "../util/platform.js";
+import type { SessionRecording } from "../../../tools/browser/network-recording-types.js";
+import { getDataDir } from "../../../util/platform.js";
 
 export type AmazonRequestKey =
   | "search"
