+++ assistant/src/cli/commands/amazon/search.ts
